new pages and dashboard
