+++ src/App.jsx
@@ -2,15 +2,19 @@ import React from 'react'
 import './App.css';
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import HomePage from './components/Homepage';
+import RequestPage from './components/RequestPage';
 import CodeInfoPage from './components/CodeInfoPage';
+import ProfilePage from './components/ProfilePage';
+import NavPage from './components/NavPage';
 
 const App = ()  => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<HomePage />}/>
-        <Route path="/code" element={<CodeInfoPage /> } />
+        <Route path="/" element={<NavPage />}/>
+        <Route path="/request" element={<RequestPage/>}/>
+        <Route path="/codeinfo" element={<CodeInfoPage /> } />
+        <Route path="/profile" element={<ProfilePage /> } />
       </Routes>
     </BrowserRouter>
   );
