+++ src/components/CodeInfoPage.css
@@ -1,0 +1,6 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+}
