+++ src/components/NavPage.css
@@ -1,0 +1,18 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+}
+
+.nav-list {
+  display: grid;
+  grid-template-columns: repeat(2, auto);
+  grid-column-gap: 5px;
+  max-width: auto;
+}
+
+.nav-list img {
+  max-width: 100%; 
+  height: auto;
+}
