+++ src/components/ProfilePage.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const ProfilePage = () => {
+  return (
+    <div>
+      <span>Profile</span>
+    </div>
+  );
+};
+
+export default ProfilePage;
