+++ src/components/RequestPage.css
@@ -1,28 +1,4 @@
-/* .logo {
-    width: 15%; 
-    height: auto;
-    margin-left: auto;
-    margin-right: auto;
-    display: inline-block
-  }
-  
-  .title {
-    font-size: 400%;
-    font-weight: bold;
-    font-family: Arial, Helvetica, sans-serif;
-    text-align: center;
-  }
-  
-  .link{
-    text-decoration: none;
-    color: inherit;
-  }
-
-  .inner{
-    text-align: center;
-  } */
-
-  .container {
+.container {
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -52,9 +28,9 @@
 }
 
 .form-container {
-    display: flex;
-    flex-direction: column;
-    width: 75%;
+    /* display: flex;
+    flex-direction: column; */
+    width: 85%;
     padding: 20px;
     background-color: #f5f5f5;
     border-radius: 10px;
