+++ src/icons/mainIcon.png
